requirements.txt
aiofiles==23.1.0
aiogram==3.1.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
Byond-API==0.2.2
byond-topic==0.1.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
frozenlist==1.4.0
google-api-core==2.14.0
google-api-python-client==2.48.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
Google-Images-Search==1.4.6
googleapis-common-protos==1.61.0
greenlet==3.0.1
httplib2==0.22.0
idna==3.4
importlib==1.0.4
magic-filter==1.0.11
multidict==6.0.4
Pillow==10.1.0
protobuf==4.25.1
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyfiglet==0.8.post1
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.1.1
pytenor==1.2
python-resize-image==1.1.20
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.23
TenGiphPy==2.5.0
termcolor==1.1.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.2
